requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
google-cloud-vision==3.7.0
python-multipart==0.0.6
gunicorn==21.2.0
pydantic>=2.0.0
